Fix auto localizing URL
Issue:
When a route of 'a/b/{c}' in en and aa/bb/{c} in vi produces a URL like
http://katniss.linhntaim.com/a/b/C?d=e#f, then its auto-localized URL
for locale vi will be http://katniss.linhntaim.com/vi/a/b/C => that's
wrong.
I just fixed it in package mcamara/laravel-localization >
LaravelLocalization.php. The result URL will be
http://katniss.linhntaim.com/vi/aa/bb/C?d=e#f => exactly.